% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiR.R
\docType{data}
\name{a2}
\alias{a2}
\title{Ouput of the 1 network for 2 case study}
\format{A data frame with rows and variables}
\description{
Ouput of the 1 network for 2 case study
}
\keyword{internal}

