% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiR.R
\docType{data}
\name{c2}
\alias{c2}
\title{Ouput of the 3 network for 2 case study}
\format{A data frame with rows and variables}
\description{
Ouput of the 3 network for 2 case study
}
\keyword{internal}

