% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Case_study1.R
\name{Case_Study1_loading_1_network}
\alias{Case_Study1_loading_1_network}
\title{Loading 1 network of shared protein domanin}
\usage{
Case_Study1_loading_1_network(species)
}
\arguments{
\item{species}{variable}
}
\value{
dataframe with interactions
}
\description{
Case_Study1_loading_1_network loads shared protein domain in HomoSapiens
}
\examples{
\dontrun{
a<-Case_Study1_loading_1_network(species)
}
}

