% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Case_study1.R
\name{Case_Study1_loading_2_network}
\alias{Case_Study1_loading_2_network}
\title{Loading 2 network of shared protein domanin}
\usage{
Case_Study1_loading_2_network(data)
}
\arguments{
\item{data}{the output of Case_Study1_loading_1_network}
}
\value{
dataframe with selected interactions
}
\description{
Case_Study1_loading_2_network loads shared protein domain in HomoSapiens including only miRNAs already found deregulated in PC,
}
\examples{
\dontrun{
b<-Case_Study1_loading_2_network(data=a)
}
}

