% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Case_study1.R
\name{Case_Study1_loading_3_network}
\alias{Case_Study1_loading_3_network}
\title{Loading 3 network of shared protein domanin}
\usage{
Case_Study1_loading_3_network(data, dataFilt, dataClin)
}
\arguments{
\item{data}{the output of Case_Study1_loading_2_network}

\item{dataFilt}{TCGA matrix}

\item{dataClin}{clinical data matrix}
}
\value{
dataframe with selected interactions
}
\description{
Case_Study1_loading_3_network loads shared protein domain in HomoSapiens including only the DEGs with a direct interaction among them
}
\examples{
\dontrun{
c<-Case_Study1_loading_3_network(data=b,dataFilt=dataFilt,dataClin=dataClin)
}
}

