% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Case_study1.R
\name{Case_Study1_loading_4_network}
\alias{Case_Study1_loading_4_network}
\title{Loading 4 network of shared protein domanin}
\usage{
Case_Study1_loading_4_network(TERZA_NET)
}
\arguments{
\item{TERZA_NET}{the output of Case_Study1_loading_3_network}
}
\value{
dataframe with selected interactions
}
\description{
Case_Study1_loading_4_network loads network community with the higher number of elements in the 3 network
}
\examples{
\dontrun{
d<-Case_Study1_loading_4_network(TERZA_NET=c)
}
}
