% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_study2.r
\name{Case_Study2_loading_1_network}
\alias{Case_Study2_loading_1_network}
\title{Loading 1 network of Protein Interactions (PI)}
\usage{
Case_Study2_loading_1_network(species)
}
\arguments{
\item{species}{variable}
}
\value{
dataframe with interactions
}
\description{
Case_Study2_loading_1_network loads PI in HomoSapiens
}
\examples{
\dontrun{
a2<-Case_Study2_loading_1_network(species)}
}
