% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_study2.r
\name{Case_Study2_loading_2_network}
\alias{Case_Study2_loading_2_network}
\title{Loading 2 network of Protein Interactions (PI) with miRNAs}
\usage{
Case_Study2_loading_2_network(data)
}
\arguments{
\item{data}{output of Case_Study2_loading_1_network}
}
\value{
dataframe with interactions
}
\description{
Case_Study2_loading_2_network loads PI in HomoSapiens with miRNAs already found as deregulated in BC (only interaction miRNA-gene)
}
\examples{
\dontrun{
b2<-Case_Study2_loading_2_network(data=a2)
}
}
