% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiR.R
\docType{data}
\name{Data_CANCER_normUQ_filt}
\alias{Data_CANCER_normUQ_filt}
\title{TCGA data for breast cancer}
\format{A data frame with rows and variables}
\description{
TCGA data for breast cancer
}
\keyword{internal}
