% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiR.R
\docType{data}
\name{dataClin}
\alias{dataClin}
\title{clinical data for prostate cancer}
\format{A data frame with rows and variables}
\description{
clinical data for prostate cancer
}
\keyword{internal}
