% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiR.R
\docType{data}
\name{dataFilt}
\alias{dataFilt}
\title{TCGA data for prostate cancer afetr quantile analysis}
\format{A data frame with rows and variables}
\description{
TCGA data for prostate cancer afetr quantile analysis
}
\keyword{internal}

