% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiR.R
\docType{data}
\name{net_shar_prot}
\alias{net_shar_prot}
\title{Output of SpidermiRquery_spec_networks}
\format{A data frame with rows and variables}
\description{
Output of SpidermiRquery_spec_networks
}
\keyword{internal}

