% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiR.R
\docType{data}
\name{out_net}
\alias{out_net}
\title{Output of SpidermiRdownload_net}
\format{A data frame with rows and variables}
\description{
Output of SpidermiRdownload_net
}
\keyword{internal}

