% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiR.R
\docType{package}
\name{SpidermiR}
\alias{SpidermiR}
\alias{SpidermiR-package}
\title{Download data}
\description{
SpidermiR allows you to Download data of samples from GeneMania
}
\details{
The functions you're likely to need from \pkg{SpidermiR} is
\code{SpidermiR}
Otherwise refer to the vignettes to see
how to format the documentation.
}

