% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRanalyze.R
\name{SpidermiRanalyze_Community_detection}
\alias{SpidermiRanalyze_Community_detection}
\title{Find community detection}
\usage{
SpidermiRanalyze_Community_detection(data, type)
}
\arguments{
\item{data}{SpidermiRanalyze_mirna_network output or SpidermiRanalyze_mirna_gene_complnet}

\item{type}{with the parameter type the user can choose the algorithm  to calculate the community structure
EB edge.betweenness.community
FC fastgreedy.community
WC walktrap.community
SC spinglass.community
LE leading.eigenvector.community
LP label.propagation.community}
}
\value{
a list of clusters with their number of genes
}
\description{
SpidermiRanalyze_Community_detection try to find dense subgraphs in directed or undirected graphs, by optimizing some criteria.
}
\examples{
miRNA_cN <-data.frame(gA=c('hsa-let-7a','hsa-miR-300'),gB=c('FOXM1','KPNA4'),stringsAsFactors=FALSE)
comm<-  SpidermiRanalyze_Community_detection(data=miRNA_cN,type="FC") 
}
