% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRanalyze.R
\name{SpidermiRanalyze_Community_detection_bi}
\alias{SpidermiRanalyze_Community_detection_bi}
\title{Community detection from biomarkers of interest}
\usage{
SpidermiRanalyze_Community_detection_bi(data, BI)
}
\arguments{
\item{data}{SpidermiRanalyze_Community_detection output}

\item{BI}{a set of biomarkers of interest}
}
\value{
a list with the cluster for each biomarkers of interest
}
\description{
SpidermiRanalyze_Community_detection_bi find the cluster with biomarkers of interest
}
\examples{
miRNA_cN <-data.frame(gA=c('hsa-let-7a','hsa-miR-300'),gB=c('FOXM1','KPNA4'),stringsAsFactors=FALSE)
comm<-  SpidermiRanalyze_Community_detection(data=miRNA_cN,type="FC") 
biomark_of_interest<-c("hsa-let-7a","CDK","FOXO1","hsa-miR-27a")
mol<-SpidermiRanalyze_Community_detection_bi(data=comm,BI=biomark_of_interest)
}
