% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRanalyze.R
\name{SpidermiRanalyze_Community_detection_net}
\alias{SpidermiRanalyze_Community_detection_net}
\title{Find the network of community detection and direct biormarker}
\usage{
SpidermiRanalyze_Community_detection_net(data, comm_det, size)
}
\arguments{
\item{data}{SpidermiRanalyze_mirna_network output or SpidermiRanalyze_mirna_gene_complnet}

\item{comm_det}{SpidermiRanalyze_Community_detection}

\item{size}{the index of community detection obtained from SpidermiRanalyze_Community_detection}
}
\value{
dataframe with the interatcions
}
\description{
SpidermiRanalyze_direct_net find the direct interactions from a specific community
}
\examples{
miRNA_cN <-data.frame(gA=c('hsa-let-7a','hsa-miR-300'),gB=c('FOXM1','KPNA4'),stringsAsFactors=FALSE)
comm<-  SpidermiRanalyze_Community_detection(data=miRNA_cN,type="FC") 
cd_net<-SpidermiRanalyze_Community_detection_net(data=miRNA_cN,comm_det=comm,size=1)
}
