% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRanalyze.R
\name{SpidermiRanalyze_degree_centrality}
\alias{SpidermiRanalyze_degree_centrality}
\title{Ranking degree centrality genes}
\usage{
SpidermiRanalyze_degree_centrality(data, cut = NULL)
}
\arguments{
\item{data}{SpidermiRanalyze_mirna_network output or SpidermiRanalyze_mirna_gene_complnet}

\item{cut}{parameter cut is able to cut off other genes}
}
\value{
dataframe with the ranked number of direct neighobors for each gene of the network
}
\description{
SpidermiRanalyze_degree_centrality provides degree centrality, defined as the total number of direct neighbors for each gene.
}
\examples{
miRNA_cN <-data.frame(gA=c('hsa-let-7a','hsa-miR-300'),gB=c('FOXM1','KPNA4'),stringsAsFactors=FALSE)
biomark_of_interest<-c("hsa-let-7a","CDK","FOXO1","hsa-miR-27a")
top10_cent<-SpidermiRanalyze_degree_centrality(miRNA_cN)
}
