% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRanalyze.R
\name{SpidermiRanalyze_direct_net}
\alias{SpidermiRanalyze_direct_net}
\title{Searching by biomarkers of interest with direct interaction}
\usage{
SpidermiRanalyze_direct_net(data, BI)
}
\arguments{
\item{data}{SpidermiRanalyze_mirna_network output or SpidermiRanalyze_mirna_gene_complnet}

\item{BI}{a set of biomarkers of interest}
}
\value{
dataframe with direct interaction of biomarkers of interest
}
\description{
SpidermiRanalyze_direct_net  finds other biomarkers that are related to a set of biomarkers of interest (the input of user) with direct interations.
}
\examples{
miRNA_cN <-data.frame(gA=c('hsa-let-7a','FOXM1'),gB=c('FOXM1','KPNA4'),stringsAsFactors=FALSE)
biomark_of_interest<-c("hsa-let-7a","CDK","FOXO1","hsa-miR-27a")
GIdirect_net<-SpidermiRanalyze_direct_net(data=miRNA_cN,BI=biomark_of_interest)
}
