% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRanalyze.R
\name{SpidermiRanalyze_mirna_extra_cir}
\alias{SpidermiRanalyze_mirna_extra_cir}
\title{Integration of Extracellular/Circulating miRNA}
\usage{
SpidermiRanalyze_mirna_extra_cir(data, type = NULL)
}
\arguments{
\item{data}{SpidermiRanalyze_mirna_gene_complnet output or network with miRNA in the first column of the dataframe}

\item{type}{The user using the following parameteres can specify the network type
\tabular{ll}{
mT \tab   to obtain a microRNA target interactions \cr
mCT  \tab   to obtain a microRNA target gene complete iteractions\cr}}
}
\value{
dataframe with Extracellular/Circulating miRNA target interaction data
}
\description{
SpidermiRanalyze_mirna_extra_cir creates a data frame with miRNA target or miRNA target gene interaction.
}
\examples{
miRNA_cN <-data.frame(gA=c('hsa-let-7a','hsa-miR-141'),gB=c('FOXM1','CDK'),stringsAsFactors=FALSE)
miRNA_NET_ext_circmT<-SpidermiRanalyze_mirna_extra_cir(data=miRNA_cN,"mT")
}

