% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRanalyze.R
\name{SpidermiRanalyze_mirna_gene_complnet}
\alias{SpidermiRanalyze_mirna_gene_complnet}
\title{Integration of microRNA target gene networks.}
\usage{
SpidermiRanalyze_mirna_gene_complnet(data, disease = NULL)
}
\arguments{
\item{data}{SpidermiRprepare_NET output}

\item{disease}{miRNA target gene interaction can be filtered by disease using the parameters obatined from SpidermiRquery_disease}
}
\value{
dataframe with miRNA target gene interaction data
}
\description{
SpidermiRanalyze_mirna_gene_complnet creates a data frame with miRNA target gene interaction. The user can filter the search by disease.
}
\examples{
GS_net <- data.frame(gA=c('SMAD','MYC'),gB=c('FOXM1','KRAS'),stringsAsFactors=FALSE)
miRNA_cNT<-SpidermiRanalyze_mirna_gene_complnet(data=GS_net,disease="prostate cancer")
}

