% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRanalyze.R
\name{SpidermiRanalyze_mirnanet_pharm}
\alias{SpidermiRanalyze_mirnanet_pharm}
\title{Integration of pharmacomiR in the network}
\usage{
SpidermiRanalyze_mirnanet_pharm(mir_ph, net)
}
\arguments{
\item{mir_ph}{SpidermiRdownload_pharmacomir output}

\item{net}{a network data (e.g. SpidermiRanalyze_mirna_network or SpidermiRanalyze_mirna_gene_complnet output)}
}
\value{
a dataframe with the integation of network and pharmacomiR data
}
\description{
SpidermiRanalyze_mirnanet_pharm integrates both miRNA targeting of the gene and the gene-drug interaction from PharmacomiR database in the network
}
\examples{
\dontrun{
mir_p <-data.frame(gA=c('hsa-let-7a','CASP3'),gB=c('CASP3','paclitaxel'),stringsAsFactors=FALSE)
net_p <-data.frame(gA=c('hsa-let-7a','hsa-miR-300'),gB=c('FOXM1','KPNA4'),stringsAsFactors=FALSE)
mol<-SpidermiRanalyze_mirnanet_pharm(mir_ph=mir_p,net=net_p)}
}
