% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRanalyze.R
\name{SpidermiRanalyze_subnetwork_neigh}
\alias{SpidermiRanalyze_subnetwork_neigh}
\title{Searching by biomarkers of interest and all the edges among this bunch of nodes}
\usage{
SpidermiRanalyze_subnetwork_neigh(data, BI)
}
\arguments{
\item{data}{SpidermiRanalyze_mirna_network output or SpidermiRanalyze_mirna_gene_complnet}

\item{BI}{a set of biomarkers of interest}
}
\value{
dataframe with interactions
}
\description{
SpidermiRanalyze_subnetwork_neigh create a sub network composed by the nodes in BI and, if some of them are connected to other nodes (even if not in BI), take also them (include all the edges among this bunch of nodes).
}
\examples{
miRNA_cN <-data.frame(gA=c('hsa-let-7a','hsa-miR-300'),gB=c('FOXM1','KPNA4'),stringsAsFactors=FALSE)
biomark_of_interest<-c("hsa-let-7a","CDK","FOXO1","hsa-miR-27a")
GIdirect_net_neigh<-SpidermiRanalyze_subnetwork_neigh(data=miRNA_cN,BI=biomark_of_interest)
}
