% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRdownload.R
\name{SpidermiRdownload_drug_gene}
\alias{SpidermiRdownload_drug_gene}
\title{Download drug-gene interactions in  DGIdb}
\usage{
SpidermiRdownload_drug_gene(drug_gene)
}
\arguments{
\item{drug_gene}{parameter}
}
\value{
a dataframe with miRNA target validated interactions
}
\description{
SpidermiRdownload_drug_gene will download drug gene interactions
}
\examples{
drug_genetarget<-SpidermiRdownload_drug_gene(drug_gene)
}
