% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRdownload.R
\name{SpidermiRdownload_miRNAextra_cir}
\alias{SpidermiRdownload_miRNAextra_cir}
\title{Download miRNA validated database}
\usage{
SpidermiRdownload_miRNAextra_cir(miRNAextra_cir)
}
\arguments{
\item{miRNAextra_cir}{parameter}
}
\value{
a dataframe with miRNA target validated interactions
}
\description{
SpidermiRdownload_miRNAprediction will download miRNA validated target
}
\examples{
list<-SpidermiRdownload_miRNAextra_cir(miRNAextra_cir)
}

