% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRdownload.R
\name{SpidermiRdownload_miRNALNC_miRNA}
\alias{SpidermiRdownload_miRNALNC_miRNA}
\title{Download microRNAs binding sites on long non coding RNA from  lncRNome database}
\usage{
SpidermiRdownload_miRNALNC_miRNA(miRNALNC_miRNA)
}
\arguments{
\item{miRNALNC_miRNA}{parameter}
}
\value{
a dataframe with miRNALNC_miRNA
}
\description{
SpidermiRdownload_miRNALNC_miRNA-RNA will download microRNAs binding sites on long non coding RNA
}
\examples{
list_LNC_miRNA<-SpidermiRdownload_miRNALNC_miRNA(miRNALNC_miRNA)
}
