% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRdownload.R
\name{SpidermiRdownload_miRNALNC_RNA}
\alias{SpidermiRdownload_miRNALNC_RNA}
\title{Download Long Non-Coding RNA (LNC-RNA) from  lncRNome database}
\usage{
SpidermiRdownload_miRNALNC_RNA(miRNALNC_RNA)
}
\arguments{
\item{miRNALNC_RNA}{parameter}
}
\value{
a dataframe with miRNALNC_RNA
}
\description{
SpidermiRdownload_LNC-RNA will download LNC-RNA
}
\examples{
list<-SpidermiRdownload_miRNALNC_RNA(miRNALNC_RNA)
}
