% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRdownload.R
\name{SpidermiRdownload_miRNAprediction}
\alias{SpidermiRdownload_miRNAprediction}
\title{Download human miRNA predicted database}
\usage{
SpidermiRdownload_miRNAprediction(mirna_list)
}
\arguments{
\item{mirna_list}{miRNA list of interest}
}
\value{
a dataframe with miRNA target validated interactions
}
\description{
SpidermiRdownload_miRNAprediction will download miRNA predicted target
}
\examples{
mirna<-c('hsa-miR-567')
list<-SpidermiRdownload_miRNAprediction(mirna_list=mirna)
}
