% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRdownload.R
\name{SpidermiRdownload_miRNAprediction_mmu}
\alias{SpidermiRdownload_miRNAprediction_mmu}
\title{Download mmu miRNA predicted database}
\usage{
SpidermiRdownload_miRNAprediction_mmu(mirna_list)
}
\arguments{
\item{mirna_list}{miRNA list of interest}
}
\value{
a dataframe with miRNA target validated interactions
}
\description{
SpidermiRdownload_miRNAprediction_mmu will download miRNA predicted target
}
\examples{
mirna<-c('mmu-miR-708-3p')
list<-SpidermiRdownload_miRNAprediction_mmu(mirna_list=mirna)
}
