% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRdownload.R
\name{SpidermiRdownload_miRNAvalidate}
\alias{SpidermiRdownload_miRNAvalidate}
\title{Download miRNA validated database}
\usage{
SpidermiRdownload_miRNAvalidate(validated)
}
\arguments{
\item{validated}{parameter}
}
\value{
a dataframe with miRNA target validated interactions
}
\description{
SpidermiRdownload_miRNAprediction will download miRNA validated target
}
\examples{
list<-SpidermiRdownload_miRNAvalidate(validated)
}
