% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRdownload.R
\name{SpidermiRdownload_net}
\alias{SpidermiRdownload_net}
\title{Download the network from GeneMania.}
\usage{
SpidermiRdownload_net(data)
}
\arguments{
\item{data}{The SpidermiRquery_spec_networks output}
}
\value{
Download GeneMania network
}
\description{
SpidermiRdownload_net function will download the data
}
\examples{
out_net<-SpidermiRdownload_net(data=net_shar_prot)
}

