% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRdownload.R
\name{SpidermiRdownload_pharmacomir}
\alias{SpidermiRdownload_pharmacomir}
\title{Download drug-miRNA interactions with fisher test}
\usage{
SpidermiRdownload_pharmacomir(list, drug)
}
\arguments{
\item{list}{miRNA gene target as obtained from e.g.; SpidermiRdownload_miRNAvalidate}

\item{drug}{parameter drug of interest}
}
\value{
a dataframe with miRNA target validated interactions
}
\description{
SpidermiRdownload_pharmacomir will calculate miRNA- drug interactions
}
\examples{
list1<-SpidermiRdownload_miRNAvalidate(validated)
drug="TAMOXIFEN"
drug_genetarget<-SpidermiRdownload_pharmacomir(list1[1:100,],drug="TAMOXIFEN")
}
