% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRquery.R
\name{SpidermiRquery_disease}
\alias{SpidermiRquery_disease}
\title{Visualize disease categories}
\usage{
SpidermiRquery_disease(diseaseID)
}
\arguments{
\item{diseaseID}{variable name}
}
\value{
a list of disease.
}
\description{
The user can visualize the disease supported by SpidermiR
}
\examples{
disease<-SpidermiRquery_disease(diseaseID)
}

