% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRquery.R
\name{SpidermiRquery_networks_type}
\alias{SpidermiRquery_networks_type}
\title{Network categories}
\usage{
SpidermiRquery_networks_type(organismID)
}
\arguments{
\item{organismID}{describes index of a specific specie obtained by SpidermiRquery_species output}
}
\value{
a list of network categories in a specie indicated.
}
\description{
The user can visualize the network types supported by GeneMania for a specific specie using SpidermiRquery_networks_type
}
\examples{
org<-SpidermiRquery_species(species)
net_type<-SpidermiRquery_networks_type(organismID=org[9,])
}
