% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRquery.R
\name{SpidermiRquery_spec_networks}
\alias{SpidermiRquery_spec_networks}
\title{Searching by network categories}
\usage{
SpidermiRquery_spec_networks(organismID, network)
}
\arguments{
\item{organismID}{describes index of a specific specie obtained by SpidermiRquery_species output}

\item{network}{The network type the user is interested in.
Example:
\tabular{ll}{
COexp \tab   Co-expression \cr
PHint  \tab   Physical_interactions \cr
COloc  \tab   Co-localization \cr
GENint  \tab   Genetic_interactions \cr
PATH  \tab   Pathway \cr
SHpd  \tab   Shared_protein_domains \cr
pred  \tab   Predicted \cr
}}
}
\value{
a list of the database or reference where the information came from.
}
\description{
The user can visualize the database or reference where the information came from
}
\examples{
org<-SpidermiRquery_species(species)
net_shar_prot<-SpidermiRquery_spec_networks(organismID = org[9,],
                                        network = "SHpd")
}
