% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRquery.R
\name{SpidermiRquery_species}
\alias{SpidermiRquery_species}
\title{Searching  by network species}
\usage{
SpidermiRquery_species(species)
}
\arguments{
\item{species}{a variable parameter}
}
\value{
table of species
}
\description{
The user can visualize the species supported by GeneMania, using the function SpidermiRquery_species .
}
\examples{
org<-SpidermiRquery_species(species)
}
