% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRvisualize.R
\name{SpidermiRvisualize_3Dbarplot}
\alias{SpidermiRvisualize_3Dbarplot}
\title{plots the 3D barplot}
\usage{
SpidermiRvisualize_3Dbarplot(Edges_1net, Edges_2net, Edges_3net, Edges_4net,
  Edges_5net, NODES_1net, NODES_2net, NODES_3net, NODES_4net, NODES_5net,
  nmiRNAs_1net, nmiRNAs_2net, nmiRNAs_3net, nmiRNAs_4net, nmiRNAs_5net)
}
\arguments{
\item{Edges_1net}{int number of edges in the 1 net}

\item{Edges_2net}{int number of edges in the 2 net}

\item{Edges_3net}{int number of edges in the 3 net}

\item{Edges_4net}{int number of edges in the 4 net}

\item{Edges_5net}{int number of edges in the 5 net}

\item{NODES_1net}{int number of nodes in the 1 net}

\item{NODES_2net}{int number of nodes in the 2 net}

\item{NODES_3net}{int number of nodes in the 3 net}

\item{NODES_4net}{int number of nodes in the 4 net}

\item{NODES_5net}{int number of nodes in the 5 net}

\item{nmiRNAs_1net}{int number of miRNAs in the 1 net}

\item{nmiRNAs_2net}{int number of miRNAs in the 2 net}

\item{nmiRNAs_3net}{int number of miRNAs in the 3 net}

\item{nmiRNAs_4net}{int number of miRNAs in the 4 net}

\item{nmiRNAs_5net}{int number of miRNAs in the 5 net}
}
\value{
barplot
}
\description{
It shows a barplot of 5 networks given by the user with a summary representation of number of nodes, edges, and miRNAs (log values)
}
\examples{
SpidermiRvisualize_3Dbarplot(Edges_1net=1041003,Edges_2net=100016,Edges_3net=3008,
Edges_4net=1493,Edges_5net=1598,NODES_1net=16502,NODES_2net=13338,NODES_3net=1429,NODES_4net=675,
NODES_5net=712,nmiRNAs_1net=0,nmiRNAs_2net=74,nmiRNAs_3net=0,nmiRNAs_4net=0,nmiRNAs_5net=37)
}
