% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRvisualize.R
\name{SpidermiRvisualize_adj_matrix}
\alias{SpidermiRvisualize_adj_matrix}
\title{plots the adjacency matrix of the network}
\usage{
SpidermiRvisualize_adj_matrix(data)
}
\arguments{
\item{data}{The input data is a network}
}
\value{
plot
}
\description{
It shows a plot OF the adjacency matrix of the network
}
\examples{
cd<-data.frame(gA=c('hsa-let-7a','hsa-miR-141'),gB=c('FOXM1','CDK'),stringsAsFactors=FALSE)
SpidermiRvisualize_adj_matrix(data=cd)
}
