% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRvisualize.R
\name{SpidermiRvisualize_BI}
\alias{SpidermiRvisualize_BI}
\title{Visualize results obtained by SpidermiR analysis starting form a set of biomarker of interest}
\usage{
SpidermiRvisualize_BI(data, BI)
}
\arguments{
\item{data}{The input data is a dataframe containing network data.}

\item{BI}{a set of biomarkers of interest}
}
\value{
3D graphic
}
\description{
Visualize miRNA-target interaction and miRNA-target-gene starting from a set of biomarker of interest
}
\examples{
miRNA_cNET <-data.frame(gA=c('hsa-let-7a','hsa-miR-141'),gB=c('FOXM1','CDK'),stringsAsFactors=FALSE)
biomark_of_interest<-c("hsa-let-7a","CDK","FOXO1","hsa-miR-27a")
SpidermiRvisualize_BI(data=miRNA_cNET,BI=biomark_of_interest)
}
