% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRvisualize.R
\name{SpidermiRvisualize_degree_dist}
\alias{SpidermiRvisualize_degree_dist}
\title{plots the degree distribution of the network}
\usage{
SpidermiRvisualize_degree_dist(data)
}
\arguments{
\item{data}{The input data is a network}
}
\value{
plot
}
\description{
It shows a plot of the degree distribution of the network
}
\examples{
cd<-data.frame(gA=c('hsa-let-7a','hsa-miR-141'),gB=c('FOXM1','CDK'),stringsAsFactors=FALSE)
SpidermiRvisualize_degree_dist(data=cd)
}
