% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRvisualize.R
\name{SpidermiRvisualize_mirnanet}
\alias{SpidermiRvisualize_mirnanet}
\title{Visualize results obtained by SpidermiR analysis}
\usage{
SpidermiRvisualize_mirnanet(data)
}
\arguments{
\item{data}{The input data is a dataframe containing network data.}
}
\value{
3D graphic
}
\description{
Visualize the network
}
\examples{
miRNA_cNET <-data.frame(gA=c('hsa-let-7a','hsa-miR-141'),gB=c('FOXM1','CDK'),stringsAsFactors=FALSE)
SpidermiRvisualize_mirnanet(data=miRNA_cNET)
}

