% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRanalyze.R
\name{SpidermiRanalyze_DEnetworkTCGA}
\alias{SpidermiRanalyze_DEnetworkTCGA}
\title{Integration with TCGA data in order to obtain a network of  differentially expressed (DE) genes or miRNAs.}
\usage{
SpidermiRanalyze_DEnetworkTCGA(data, TCGAmatrix, tumour, normal)
}
\arguments{
\item{data}{network data (e.g. shared protein domains, co-expression,..)}

\item{TCGAmatrix}{gene or miRNA expression matrix}

\item{tumour}{barcode TCGA tumour data}

\item{normal}{barcode TCGA normal data}
}
\value{
a network miRNA-gene differentially expressed as calculated by TCGAbiolinks package. The user can select the samples and cancer type from TCGA portal.
}
\description{
SpidermiRanalyze_DEnetworkTCGA integrates the information of differential analysis of TCGA data in the network. The final result will be a network with only DE genes or miRNAs depending  whether the user chooses to mRNA or miRNA  TCGA data.
}
\examples{
miRNA_cN <-data.frame(gA=c('IGFL3','GABRA1'),gB=c('IGFL2','KRT13'),stringsAsFactors=FALSE)
tumour<-c("TCGA-E9-A1RD-11A","TCGA-E9-A1RC-01A")
normal<-c("TCGA-BH-A18P-11A","TCGA-BH-A18L-11A") 
de_int<-SpidermiRanalyze_DEnetworkTCGA(data=miRNA_cN,
                                       TCGAmatrix=Data_CANCER_normUQ_filt,
                                       tumour,
                                       normal
                                      )
}

