% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpidermiRdownload.R
\name{SpidermiRdownload_pharmacomir}
\alias{SpidermiRdownload_pharmacomir}
\title{Download both miRNA target and the gene-drug interaction from PharmacomiR database}
\usage{
SpidermiRdownload_pharmacomir(pharmacomir)
}
\arguments{
\item{pharmacomir}{variable}
}
\value{
a dataframe with gene-drug, and miR-gene associations
}
\description{
SpidermiRdownload_pharmacomir will download miRNA Pharmacogenomic data
}
\examples{
\dontrun{
mir_pharmaco<-SpidermiRdownload_pharmacomir(pharmacomir=pharmacomir)}
}
